+++ app/renderer/components/common/commonForm.js
@@ -3,14 +3,14 @@
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 const React = require('react')
-const ImmutableComponent = require('./immutableComponent')
+const ImmutableComponent = require('../immutableComponent')
 
 const {StyleSheet, css} = require('aphrodite/no-important')
-const globalStyles = require('./styles/global')
-const commonStyles = require('./styles/commonStyles')
+const globalStyles = require('../styles/global')
+const commonStyles = require('../styles/commonStyles')
 
-const {FormDropdown} = require('./dropdown')
-const {FormTextbox} = require('./textbox')
+const {FormDropdown} = require('../dropdown')
+const {FormTextbox} = require('../textbox')
 
 class CommonForm extends ImmutableComponent {
   render () {
